Rework plugin to use auth_aclcheck()
This replaces the code for checking the ACL rules by calls to the
existing auth_aclcheck() function. This avoids the duplication of
ACL evaluation logic and allows to use this plugin with ACL
extenstion plugins like aclregex or aclplusregex.

The new logic goes through all the users and groups that are
contained in the ACL rules and checks the permissions for these
subjects. If the permissions for the subject exceed those of
the 'ALL' group, they are added to the list.

Signed-off-by: Frieder Schrempf <dev@fris.de>